backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.2.0
python-pptx==0.6.23
groq==1.0.0
python-multipart==0.0.9
python-dotenv==1.0.1
openpyxl==3.1.2
